Fix deserialization bug
- This fixes deserialization failure when nested objects are null.
- This also addes:
   - IntegerArchivable as a generic archivable type
   - ArchiverObjectScope to manage ObjectArchivable using RAII pattern

+++ src/core/modules/serviceworker/ServiceWorkerUpdateViaCache.h
@@ -24,7 +24,7 @@ namespace Starfish {
 
 class String;
 
-enum class ServiceWorkerUpdateViaCache : uint8_t {
+enum class ServiceWorkerUpdateViaCache : unsigned {
     Imports,
     All,
     None,
